Revert "終極改動 3000行土法煉鋼 評論瀏覽"
This reverts commit 5a17529c4306a2191518c7df0167f4f66f5ca3b3.

+++ MaiMai/MaiMai/Models/Comment.cs
@@ -16,7 +16,7 @@ namespace MaiMai.Models
     {
         public int CommentID { get; set; }
         public Nullable<int> OrderID { get; set; }
-        public Nullable<double> starRate { get; set; }
+        public Nullable<int> starRate { get; set; }
         public string commentDescription { get; set; }
         public int UserID { get; set; }
         public Nullable<int> CommentorUserID { get; set; }
